[VSPC-150] Changes for review comments related to slide.html changes, ReferenceBlock, private variables in class, reference display interface
