+++ vspace/src/main/java/edu/asu/diging/vspace/web/staff/AddReferenceController.java
@@ -15,10 +15,10 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import edu.asu.diging.vspace.core.model.IBiblioBlock;
 import edu.asu.diging.vspace.core.model.IReference;
 import edu.asu.diging.vspace.core.model.impl.Reference;
-import edu.asu.diging.vspace.core.model.impl.ReferenceBlock;
 import edu.asu.diging.vspace.core.services.IContentBlockManager;
 import edu.asu.diging.vspace.core.services.IReferenceManager;
-import edu.asu.diging.vspace.references.ReferenceDisplayProvider;
+import edu.asu.diging.vspace.references.ReferenceBlock;
+import edu.asu.diging.vspace.references.ReferenceDisplayDefault;
 
 @Controller
 public class AddReferenceController {
@@ -30,7 +30,7 @@ public class AddReferenceController {
     private IContentBlockManager contentBlockManager;
     
     @Autowired
-    private ReferenceDisplayProvider referenceDisplayProvider;
+    private ReferenceDisplayDefault referenceDisplayProvider;
 
     @RequestMapping(value = "/staff/module/{id}/slide/{slideId}/biblio/{biblioId}/reference/add", method = RequestMethod.POST)
     public ResponseEntity<ReferenceBlock> addReference(@PathVariable("id") String moduleId, @PathVariable("slideId") String slideId,
